% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\docType{data}
\name{invert}
\alias{invert}
\title{Macroinvertebrates of Loire River data}
\format{A list containing:\cr
    - \code{spe} 38 observations of 40 invertebrate species,\cr
    - \code{env} 38 observations of 3 aquatic enviro variables,\cr
    - \code{tra} 11 traits for 40 species,\cr
    - \code{phy} phylogeny for 40 species.}
\source{
Ivol et al. (1997).
}
\usage{
invert
}
\description{
This data set gives information about invertebrate species,
    environment, traits, and phylogeny.
}
\details{
See \code{?macroloire} in \code{ade4}.
}
\examples{
\dontrun{?macroloire} # in package 'ade4'
data(invert)

}
\references{
Ivol, J. M., Guinand, B., Richoux, P. and Tachet, H. 1997.
     Longitudinal changes in Trichoptera and Coleoptera assemblages
     and environmental conditions in the Loire River (France).
     Archiv for Hydrobiologie 138: 525–557.
}
\keyword{datasets}
