% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordfn.R
\name{ordfn}
\alias{ordfn}
\title{Nonmetric multidimensional scaling helper}
\usage{
ordfn(spe, distance, k, ...)
}
\arguments{
\item{spe}{species abundance matrix}

\item{distance}{dissimilarity index, per \code{vegdist}}

\item{k}{number of ordination axes sought}

\item{...}{additional arguments passed to \code{metaMDS}}
}
\value{
Object of class \code{metaMDS}.
}
\description{
Light wrapper around \code{metaMDS} with some
    pre-specified arguments.
}
\details{
Very light wrapper around \code{metaMDS}.
}
\examples{
data(veg) # load('./data/veg.rda', verbose=T)
spe <- veg$spe
(m <- ordfn(spe, dist='bray', k=2))
plot(m)

}
\seealso{
\code{metaMDS}
}
