% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persp_heat.R
\name{persp_heat}
\alias{persp_heat}
\title{Add heatmap colors to a perspective plot}
\usage{
persp_heat(z, pal = "heat", ...)
}
\arguments{
\item{z}{matrix of values}

\item{pal}{character string, one of \code{'heat'} for heat colors,
or \code{'redblue'} for diverging red-blue palette}

\item{...}{additional arguments passed to \code{persp}}
}
\value{
\code{persp} plot.
}
\description{
Helper function to add heatmap colors to a perspective
    plot.
}
\details{
Plots a perspective plot with facets colored by z-value.
}
\examples{
# from the persp documentation:
x <- seq(-10, 10, length= 30)
y <- x
f <- function(x, y) { r <- sqrt(x^2+y^2); 10 * sin(r)/r }
z <- outer(x, y, f)
persp(x, y, z, theta = 30, phi = 30, expand = 0.5, col = 'lightblue')
persp_heat(z, theta = 30, phi = 30, expand = 0.5)
persp_heat(z, pal='redblue', theta = 30, phi = 30, expand = 0.5)

}
\seealso{
\code{\link[vegan]{ordiplot}}
}
