% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_corr.R
\name{phylo_corr}
\alias{phylo_corr}
\title{Phylogenetic correction of traits}
\usage{
phylo_corr(phy, tra, ...)
}
\arguments{
\item{phy}{phylogenetic tree of class \code{'phylo'}}

\item{tra}{matrix or data.frame, rows=species and cols=traits}

\item{...}{additional arguments (currently ignored)}
}
\value{
Data frame of same dimensions as \code{tra}, but with
    phylogenetically-corrected values.
}
\description{
Remove potential phylogenetic signal from a set of
    continuous or discrete traits.
}
\details{
Removes phylogenetic signal from trait values, by applying
    correction factors obtained from Cholesky decomposition of the
    phylogenetic variance-covariance matrix (Butler et al. 2000).
    Current implementation follows Eklöf and Stouffer (2016) in
    permitting both continuous and categorical traits.
}
\examples{
data(veg) # load('./data/veg.rda', verbose=T)
phy <- veg$phy
tra <- veg$tra
ptra <- phylo_corr(phy, tra)
ecole::set_par(8)
for(i in 1:8){
     plot(tra[,i], ptra[,i], pch=16, cex=0.7, col='#00000050',
     xlab=dimnames(tra)[[2]][i], ylab=dimnames(ptra)[[2]][i])
}

}
\references{
Butler, M.A., T.W. Schoener, and J.B. Losos. 2000. The relationship
    between sexual size dimorphism and habitat use in Greater
    Antillean Anolis lizards. Evolution 54:259-272.\cr
Eklöf, A., and D.B. Stouffer. 2016. The phylogenetic component of
    food web structure and intervality. Theoretical Ecology
    9:107–115.\cr
}
\seealso{
Code adapted from: https://github.com/stoufferlab/phyloint/
}
