% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wa.R
\name{plot_wa}
\alias{plot_wa}
\title{Plot centroids and ranges}
\usage{
plot_wa(spe, mod, xexp = 1.6, yexp = 1, pick = NULL, lcol,
  buff = 0.1, ...)
}
\arguments{
\item{spe}{species abundance matrix used in \code{mod}}

\item{mod}{ordination model from \code{vegan}}

\item{xexp, yexp}{numeric, expansion factor making room for x,y
labels}

\item{pick}{numeric, single ordination axis to plot (default =
\code{NULL} will plot first 2 axes as a biplot)}

\item{lcol}{color for range lines}

\item{buff}{numeric, expansion factor giving buffer around range
lines}

\item{...}{additional arguments passed to
\code{\link[vegan]{ordiplot}}}
}
\value{
Plots to device.
}
\description{
Plot species centroids and ranges in ordinations.
}
\details{
Plot species centroids and ranges in ordination spaces or single
    ordination axes.
}
\examples{
data(veg) # load('./data/veg.rda', verbose=T)
spe <- veg$spe
m <- vegan::metaMDS(vegan::vegdist(spe), trace=0)
plot_wa(spe, m, cex=0.7, las=1)

}
\seealso{
\code{\link[vegan]{ordiplot}}
}
