% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg.R
\docType{data}
\name{veg}
\alias{veg}
\title{Mafragh, Algeria vegetation data}
\format{A list containing:\cr
    - \code{xy}  97 observations of 2 spatial coordinates,\cr
    - \code{spe} 97 observations of 56 plant species,\cr
    - \code{env} 347 observations of 11 soil enviro variables,\cr
    - \code{tra} 12 traits for 56 plant species,\cr
    - \code{phy} phylogeny for 56 plant species.}
\source{
Pavoine et al. (2011).
}
\usage{
veg
}
\description{
This data set gives information about spatial coordinates, species,
    environment, traits, and phylogeny.
}
\details{
See \code{?mafragh} in \code{ade4}.
}
\examples{
\dontrun{?mafragh} # in package 'ade4'
data(veg)

}
\references{
Pavoine, S., Vela, E., Gachet, S., de Bélair, G. and Bonsall, M. B.
     2011. Linking patterns in phylogeny, traits, abiotic variables
     and space: a novel approach to linking environmental filtering
     and plant community assembly. Journal of Ecology 99:165–175.
     doi:10.1111/j.1365-2745.2010.01743.x
}
\keyword{datasets}
