% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamfit.R
\name{gamfit}
\alias{gamfit}
\alias{print.gamfit}
\alias{summary.gamfit}
\alias{plot.gamfit}
\title{GAM for environment-ordination fit}
\usage{
gamfit(ord, env, ...)

\method{print}{gamfit}(x, ...)

\method{summary}{gamfit}(object, ...)

\method{plot}{gamfit}(x, pick = 1, pcol, lcol, pcex, lwd, title = TRUE,
  ...)
}
\arguments{
\item{ord}{ordination model from \code{vegan}}

\item{env}{environmental matrix with variables to evaluate}

\item{...}{additional arguments passed to
\code{\link[mgcv]{gam}}}

\item{x, object}{object of class \code{gamfit}}

\item{pick}{numeric, the column number of environmental variable to
plot as a nonlinear regression surface in the ordination space}

\item{pcol, lcol}{point and line colors}

\item{pcex, lwd}{point size and line width}

\item{title}{logical, add variable name as title? (default=TRUE)}
}
\value{
Object of class \code{gamfit}.
}
\description{
Generalized Additive Model (GAM) to regress one or
    many environmental variables on NMDS ordination scores.
}
\details{
Nonlinear regression of each environmental variable in turn on the
    site scores resulting from NMDS ordination.  Returns p-values
    for the joint smooth term, and adjusted-R2 expressing the
    strength of (nonlinear) relationship between each environmental
    variable and the ordination. Print, summary and plot methods
    exist. Sensitivity analysis is yet to do.
}
\examples{
data(veg) # load('./data/veg.rda', verbose=T)
spe <- veg$spe
env <- veg$env
m <- vegan::metaMDS(vegan::vegdist(spe), trace=0)
g <- gamfit(m, env)
g
plot(g, pick=10)

}
\seealso{
\code{\link[mgcv]{gam}} for the internal fitting function,
    \code{\link[vegan]{ordisurf}} for a similar procedure that only
    admits variables one-at-a-time, and \code{\link[vegan]{envfit}}
    for a linear alternative.
}
